# services/api/internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal OS API",
    "version": "0.7.0",
    "description": "Local-first personal platform. Phase 9: agentic depth — goals, subscription detection, transfer pairing, recurring tasks, habit schedules + rolling consistency, water intake, exercise PRs, reading highlights, expiry surfacing, live activity pulse."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "local dev" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" }
    },
...
  },
  "paths": {
...
    "/v1/goals": {
      "get": {
        "operationId": "listGoals",
        "summary": "Goals (kind=savings|calorie; calorie is single-row upsert)",
        "parameters": [{ "name": "kind", "in": "query", "schema": { "type": "string", "enum": ["savings", "calorie"] } }],
        "responses": { "200": { "description": "Goal[]" } }
      },
      "post": {
        "operationId": "createGoal",
        "summary": "Create goal. savings: {name, target_minor, deadline?}. calorie: {name, target_minor=kcal} replaces the previous.",
        "responses": { "201": { "description": "created" }, "400": { "description": "validation error" } }
      }
    },
    "/v1/goals/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": { "operationId": "getGoal", "summary": "Get goal", "responses": { "200": { "description": "ok" }, "404": { "description": "missing" } } },
      "patch": { "operationId": "updateGoal", "summary": "Update name/target/saved/deadline", "responses": { "200": { "description": "ok" } } },
      "delete": { "operationId": "deleteGoal", "summary": "Delete goal", "responses": { "204": { "description": "deleted" } } }
    },
...
    "/v1/items/expiring": {
      "get": {
        "operationId": "expiringItems",
        "summary": "Items whose data JSON carries a date-like field (expires/due/…) within N days",
        "parameters": [{ "name": "days", "in": "query", "schema": { "type": "integer", "default": 30, "maximum": 365 } }],
        "responses": { "200": { "description": "ExpiringItem[]" } }
      }
    },
...
  }
}